requirements.txt
Django==5.0.6
matplotlib==3.9.0
numpy==2.0.0
opencv_python==4.10.0.84
Pillow==10.4.0
Requests==2.32.3
tensorflow==2.15.0
tensorflow_intel==2.15.0
